simply Dataverse content provider (support datasets only) #1388
When the Dataverse content provider was added in #739 it had the
flexibility to operate directly on Dataverse files like this:

repo2docker https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/6ZXAGT/3YRRYJ

However, being able to operate only on datasets (files are stored
in datasets in Dataverse) is enough. That is, this will still work:

repo2docker doi:10.7910/DVN/TJCLKP

And that's all we need.

This simplification builds upon the work in #1388 where the content
of the dataset landing page is not retrieved from the DOI of the
dataset. Instead, the redirect location is fetched, which is all
the Dataverse content provider needs to determine which of the
100+ installations of Dataverse hosts the DOI.

This change should be a no-op for any installation of Datavese with
Binder integration enabled.

Harvard Dataverse (one of the 100+ installations) specifically is
not working with Binder due to a firewall that is blocking
https://dataverse.harvard.edu/citation
The simplification in this commit means that the Dataverse
content provider no longer needs to follow `/citation` to determine
what is on the other side (dataset.xhtml, file.xhtml, etc.). It
assumes that the DOI is always for a dataset (not a file), which
is the expectation we have always set for the Binder tool.

We are tracking Binder not working with Harvard Dataverse here:
https://github.com/IQSS/dataverse.harvard.edu/issues/328